% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_tuning.R
\name{plot.spaCraftTuning}
\alias{plot.spaCraftTuning}
\title{Plot Method for spaCraftTuning}
\usage{
\method{plot}{spaCraftTuning}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spaCraftTuning}.}

\item{...}{Additional arguments to be passed to plotting functions.}
}
\description{
Plot Method for spaCraftTuning
}
