% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createspaCraftObject.R
\name{testingResult}
\alias{testingResult}
\alias{testingResult,spaCraft-method}
\alias{testingResult<-}
\alias{testingResult<-,spaCraft-method}
\title{Get or Set Testing Results}
\usage{
testingResult(object)

\S4method{testingResult}{spaCraft}(object)

testingResult(object) <- value

\S4method{testingResult}{spaCraft}(object) <- value
}
\arguments{
\item{object}{A \code{spaCraft} object.}

\item{value}{A named list of endpoint results to store.}
}
\value{
\code{testingResult} returns the stored list; the replacement method
returns the updated \code{spaCraft} object.
}
\description{
Accessor and replacement methods for the \code{testing_result}
slot of a \code{spaCraft} object.
}
